+++ src/main/java/com/atlauncher/Gsons.java
@@ -18,6 +18,7 @@
 package com.atlauncher;
 
 import java.awt.Color;
+import java.time.Duration;
 import java.time.Instant;
 import java.util.Date;
 
@@ -26,6 +27,7 @@ import com.atlauncher.data.AbstractAccount;
 import com.atlauncher.data.AccountTypeAdapter;
 import com.atlauncher.data.ColorTypeAdapter;
 import com.atlauncher.data.DateTypeAdapter;
+import com.atlauncher.data.DurationTypeAdapter;
 import com.atlauncher.data.InstantTypeAdapter;
 import com.atlauncher.data.PackVersion;
 import com.atlauncher.data.PackVersionTypeAdapter;
@@ -71,6 +73,8 @@ public final class Gsons {
             .registerTypeAdapter(AbstractAccount.class, new AccountTypeAdapter())
             .registerTypeAdapter(Date.class, new DateTypeAdapter())
             .registerTypeAdapter(Color.class, new ColorTypeAdapter())
+            .registerTypeAdapter(Instant.class, new InstantTypeAdapter())
+            .registerTypeAdapter(Duration.class, new DurationTypeAdapter())
             .registerTypeAdapter(OauthTokenResponse.class, new OauthTokenResponseTypeAdapter())
             .addSerializationExclusionStrategy(exclusionAnnotationStrategy).create();
 
@@ -84,6 +88,7 @@ public final class Gsons {
 
     public static final Gson MINECRAFT = new GsonBuilder().disableHtmlEscaping().setPrettyPrinting()
             .registerTypeAdapter(Instant.class, new InstantTypeAdapter())
+            .registerTypeAdapter(Duration.class, new DurationTypeAdapter())
             .registerTypeAdapter(Color.class, new ColorTypeAdapter())
             .registerTypeAdapter(Library.class, new LibraryTypeAdapter())
             .registerTypeAdapter(Arguments.class, new ArgumentsTypeAdapter())