+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 ### New Features
 - Add dialog to prompt exe users to download bundled JRE or update it if out of date [#601]
 - New Edit Instance screen for easier, better and more thorough management of instances [#556]
+- Add createdAt and updatedAt times as well as totalPlayTime to instances
 
 ### Fixes
 - Embedded relationship mods showing as required by the launcher